Address non-sequential "-sc#" input flags
Allows for flagstring inputs which don't sequentially define the characters in the flagstring starting with -sc1, -sc2, etc. It is possible/easy to generate a single character flagstring which defines that character by the -sc4 flag while not defining and of the previous starting character flags. Existing code cannot handle this situation. Fix is to address this.