Fusion logic: Sector1.py Tourian Exit fix
Moved it to left side of acid bath, accounting for needing ball jump and freezing first ripper from right to left, and avoiding acid via SJ in both directions.

I imagine this may make future frozen enemy climbing up pirate shaft a total mess, but let's be real, adding more tricks like that in general is gonna make the entire database a complete mess. There aren't gonna be that many really dense sections like that. Bring it on.

TBH, should probably move the level 4 etanks to this connection, too, but eh, redundant and messy enough as is, and frankly, the energy requirement is just gonna be a big future system overhaul project in itself. Even RDV is lacking in proper damage-run handling.